Added a comment on the domain-joiner account format
Providing a domain user in the format <username> will fail to domain join, even if the domain is given in the domainToJoin variable.

It's important to mention this because in Terraform I see the <username> format most of the time and I've been struggling to identify why the domain join was failing (i.e. because we need username@domain.xyz format)

Thanks for the great work Paul!